Added ability to edit and delete project tasks

+++ flask_app/pflask01.py
@@ -119,47 +119,52 @@ def project_overview(project_id):
     if session.get('user'):
         p_id = project_id
         a_project = db.session.query(Project).filter_by(id=p_id).one()
-        conn = sqlite3.connect('instance/flask_note_app.db')
-        c = conn.cursor()
-        c.execute(f'SELECT * FROM tasks WHERE project_id = {p_id}')
-        return render_template('project_overview.html', project=a_project, rows=c.fetchall(), user=session['user'])
+        project_tasks = db.session.query(Task).filter_by(project_id=p_id).all()
+
+        return render_template('project_overview.html', project=a_project, tasks=project_tasks, user=session['user'])
     else:
         return redirect(url_for('login'))
 
 
 # - Edit Project -
-@app.route('/projects/edit/<project_id>', methods=['GET', 'POST'])
+@app.route('/projects/<project_id>/edit', methods=['GET', 'POST'])
 def update_project(project_id):
-    # check method used for request
-    if request.method == 'POST':
-        # get title data
-        title = request.form['title']
+    if session.get('user'):
+        # check method used for request
+        if request.method == 'POST':
+            # get title data
+            title = request.form['title']
+            a_project = db.session.query(Project).filter_by(id=project_id).one()
+            # update project data
+            a_project.title = title
+            # update project in DB
+            db.session.add(a_project)
+            db.session.commit()
+
+            return redirect(url_for('projects_list'))
+        else:
+            # GET request - show create project form to edit project
+
+            # retrieve project from database
+            a_project = db.session.query(Project).filter_by(id=project_id).one()
+
+            return render_template('project_create.html', project=a_project, user=session['user'])
+    else:
+        return redirect(url_for('login'))
+
+
+# - Delete Project -
+@app.route('/projects/<project_id>/delete', methods=['POST'])
+def delete_project(project_id):
+    if session.get('user'):
+        # retrieve project from database
         a_project = db.session.query(Project).filter_by(id=project_id).one()
-        # update project data
-        a_project.title = title
-        # update project in DB
-        db.session.add(a_project)
+        db.session.delete(a_project)
         db.session.commit()
 
         return redirect(url_for('projects_list'))
     else:
-        # GET request - show create project form to edit project
-
-        # retrieve project from database
-        a_project = db.session.query(Project).filter_by(id=project_id).one()
-
-        return render_template('project_create.html', project=a_project, user=session['user'])
-
-
-# - Delete Project -
-@app.route('/projects/delete/<project_id>', methods=['POST'])
-def delete_project(project_id):
-    # retrieve project from database
-    a_project = db.session.query(Project).filter_by(id=project_id).one()
-    db.session.delete(a_project)
-    db.session.commit()
-
-    return redirect(url_for('projects_list'))
+        return redirect(url_for('login'))
 
 
 # - Create Task for Project
@@ -183,8 +188,53 @@ def add_task(project_id):
             # ready to render response - redirect to projects list
             return redirect(url_for('project_overview', project_id=project_id))
         else:
-            # GET request - show 'create project' form
+            # GET request - show 'add task' form
             return render_template('project_task_add.html', user=session['user'])
+    else:
+        return redirect(url_for('login'))
+
+
+# - Edit Project Task -
+@app.route('/projects/<project_id>/<task_id>/edit', methods=['GET', 'POST'])
+def update_task(project_id, task_id):
+    if session.get('user'):
+        # check method used for request
+        if request.method == 'POST':
+            # get title data
+            title = request.form['title']
+            # get description data
+            description = request.form['description']
+            a_task = db.session.query(Task).filter_by(id=task_id).one()
+            # update task data
+            a_task.title = title
+            # update task description
+            a_task.description = description
+            # update task in DB
+            db.session.add(a_task)
+            db.session.commit()
+
+            return redirect(url_for('project_overview', project_id=project_id))
+        else:
+            # GET request - show create project form to edit project
+
+            # retrieve task from database
+            a_task = db.session.query(Task).filter_by(id=task_id).one()
+
+            return render_template('project_task_add.html', task=a_task, user=session['user'])
+    else:
+        return redirect(url_for('login'))
+
+
+# - Delete Task -
+@app.route('/projects/<project_id>/<task_id>/delete', methods=['POST'])
+def delete_task(project_id, task_id):
+    if session.get('user'):
+        # retrieve task from database
+        a_task = db.session.query(Task).filter_by(id=task_id).one()
+        db.session.delete(a_task)
+        db.session.commit()
+
+        return redirect(url_for('project_overview', project_id=project_id))
     else:
         return redirect(url_for('login'))
